pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crosspost"
version = "0.1.0"
description = "Post to Mastodon & Bluesky simultaneously"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Keith Lawson"}]
requires-python = ">=3.7"
dependencies = [
    "Mastodon.py>=1.8.0",
    "atproto>=0.0.40",
    "tomli>=1.1.0; python_version < '3.11'",
]
keywords = ["mastodon", "bluesky", "social-media", "crosspost"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/j-klawson/crosspost"
Issues = "https://github.com/j-klawson/crosspost/issues"

[project.scripts]
crosspost = "crosspost.cli:main"

[tool.setuptools]
packages = ["crosspost"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.data-files]
"share/man/man1" = ["man/crosspost.1"]
